+++ tina-http/src/http2/mod.rs
@@ -15,9 +15,10 @@ mod server;
 mod target;
 
 pub use client::{
-    Http2ClientConnection, Http2ClientLimits, Http2ClientMsg, Http2ClientOutcome, Http2ClientReply,
-    Http2ClientReport, Http2ClientRequest, Http2ClientRequestBody, Http2ClientResponse,
-    Http2ClientStreamCall, Http2ClientStreamingRequest, Http2ResponseChunk,
+    Http2ClientConnection, Http2ClientGrpcUnaryRequest, Http2ClientLimits, Http2ClientMsg,
+    Http2ClientOutcome, Http2ClientReply, Http2ClientReport, Http2ClientRequest,
+    Http2ClientRequestBody, Http2ClientResponse, Http2ClientStreamCall,
+    Http2ClientStreamingRequest, Http2ResponseChunk,
 };
 pub use errors::Http2ProtocolError;
 pub use server::{